Module math_vectors::vectors::vector

Structs

Vector2

Constants

DOWN
Shorthand for writing Vector2{ x: 0.0, y: -1.0 }

LEFT
Shorthand for writing Vector2{ x: -1.0, y: 0.0 }

ONE
Shorthand for writing Vector2{ x: 1.0, y: 1.0 }

RIGHT
Shorthand for writing Vector2{ x: 1.0, y: 0.0 }

UP
Shorthand for writing Vector2{ x: 0.0, y: 1.0 }

ZERO
Shorthand for writing Vector2{ x: 0.0, y: 0.0 }

Functions

are_codirectional
Return true if vectors A and B are codirectional.

are_orthogonal
Return true if vectors A and B are orthogonal. Two vectors a and b are said to be orthogonal if their direction are right angles; that's, the scalar product a * b is 0

are_parallel
Two vectors are said to be parallel if they have the same or opposite direction or if one of them is the zero vector

distance
Returns the distance between a and b.

distance_squared
Returns the squared distance between a and b.

dot
Dot Product of two vectors.

float_clamp
Clamps a value between a minimum float and maximum float value

float_clamp01
Clamps a value between 0 and 1 and returns value

lerp
Linearly interpolates between vector a and vector b by t when t = 1 returns b when t = 0 returns a when t = 0.5 returns midpoint

lerp_unclamped
Linearly interpolates between vector a and vector b by t

max
Returns a vector that is made from the largest components of two vectors

min
Returns a vector that is made from the smallest components of two vectors

reflect
Reflects a vector off the vector defined by a normal.

scale
Multiplies two vectors component-wise